Create 01.Guinea_Pig
On the first line, you will receive the quantity of food, which Merry buys for a month (30 days). On the second line, you will receive quantity hay, and on the third line – quantity cover, also for a month. On the fourth line, you will receive the guinea pig's weight.  All values are in kilograms.
Calculate whether the quantity of food, hay, and cover, will be enough for a month. The amount of food given to the guinea pig each day is 300 gr. Every second day Merry gives hay, which is 5% of the rest of the food. On every third day, Merry put cover, which is 1/3 of the guinea pig's weight. 
Puppy eats food every single day!
If food, hay, or cover runs out, stop the program.
Input:
•	On the first line – quantity food in kilograms - a floating-point number in the range [0.0 – 10000.0]
•	On the second line – quantity hay in kilograms - a floating-point number in the range [0.0 – 10000.0]
•	On the third line – quantity cover in kilograms - a floating-point number in the range [0.0 – 10000.0]
•	On the fourth line – guinea's weight in kilograms - a floating-point number in the range [0.0 – 10000.0]
Output:
•	If the food, the hay, and the cover are enough, print:
o	"Everything is fine! Puppy is happy! Food: {excessFood}, Hay: {excess hay}, Cover: {excessCover}."
•	If one of the things is not enough, print:
o	"Merry must go to the pet store!"
The output must be formatted to the second decimal place!